# apps/nextjs/lib/AgentProtocol/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agent Protocol",
    "description": "Specification of the API protocol for communication with an agent.",
    "version": "v1"
  },
  "servers": [
    { "url": "http://0.0.0.0:8000", "description": "Agent Protocol API" }
  ],
  "paths": {
...
    "/ap/v1/agent/tasks/{task_id}/artifacts": {
      "get": {
        "operationId": "listAgentTaskArtifacts",
        "summary": "List all artifacts that have been created for the given task.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "ID of the task",
            "required": true,
            "schema": { "type": "string" },
            "example": "50da533e-3904-4401-8a07-c49adf88b5eb",
            "x-postman-variables": [{ "type": "load", "name": "task_id" }]
          },
          {
            "name": "current_page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1,
              "minimum": 1
            },
            "example": 2
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "minimum": 1
            },
            "example": 25
          }
        ],
        "responses": {
          "200": {
            "description": "Returned the list of artifacts for the task.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifacts": {
                      "type": "array",
                      "items": {
                        "description": "An Artifact either created by or submitted to the agent.",
                        "type": "object",
                        "properties": {
                          "artifact_id": {
                            "description": "ID of the artifact.",
                            "type": "string",
                            "example": "b225e278-8b4c-4f99-a696-8facf19f0e56"
                          },
                          "agent_created": {
                            "description": "Whether the artifact has been created by the agent.",
                            "type": "boolean",
                            "example": false
                          },
                          "file_name": {
                            "description": "Filename of the artifact.",
                            "type": "string",
                            "example": "main.py"
                          },
                          "relative_path": {
                            "description": "Relative path of the artifact in the agent's workspace.",
                            "type": "string",
                            "example": "python/code/",
                            "nullable": true
                          }
                        },
                        "required": [
                          "artifact_id",
                          "agent_created",
                          "file_name"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total_items": {
                          "description": "Total number of items.",
                          "type": "integer",
                          "example": 42
                        },
                        "total_pages": {
                          "description": "Total number of pages.",
                          "type": "integer",
                          "example": 97
                        },
                        "current_page": {
                          "description": "Current_page page number.",
                          "type": "integer",
                          "example": 1
                        },
                        "page_size": {
                          "description": "Number of items per page.",
                          "type": "integer",
                          "example": 25
                        }
                      },
                      "required": [
                        "total_items",
                        "total_pages",
                        "current_page",
                        "page_size"
                      ]
                    }
                  },
                  "required": ["artifacts", "pagination"]
                }
              }
            }
          },
          "404": {
            "description": "Unable to find entity with a given identifier",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "description": "Message stating the entity was not found",
                      "type": "string",
                      "example": "Unable to find entity with the provided id"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "default": { "description": "Internal Server Error" }
        },
        "tags": ["agent"]
      },
      "post": {
        "operationId": "uploadAgentTaskArtifacts",
        "summary": "Upload an artifact for the specified task.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "ID of the task",
            "required": true,
            "schema": { "type": "string" },
            "example": "50da533e-3904-4401-8a07-c49adf88b5eb",
            "x-postman-variables": [{ "type": "load", "name": "task_id" }]
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "description": "Artifact to upload to the agent.",
                "type": "object",
                "properties": {
                  "file": {
                    "description": "File to upload.",
                    "type": "string",
                    "format": "binary",
                    "example": "binary representation of file"
                  },
                  "relative_path": {
                    "description": "Relative path of the artifact in the agent's workspace.",
                    "type": "string",
                    "example": "python/code"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returned the content of the artifact.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "An Artifact either created by or submitted to the agent.",
                  "type": "object",
                  "properties": {
                    "artifact_id": {
                      "description": "ID of the artifact.",
                      "type": "string",
                      "example": "b225e278-8b4c-4f99-a696-8facf19f0e56"
                    },
                    "agent_created": {
                      "description": "Whether the artifact has been created by the agent.",
                      "type": "boolean",
                      "example": false
                    },
                    "file_name": {
                      "description": "Filename of the artifact.",
                      "type": "string",
                      "example": "main.py"
                    },
                    "relative_path": {
                      "description": "Relative path of the artifact in the agent's workspace.",
                      "type": "string",
                      "example": "python/code/",
                      "nullable": true
                    }
                  },
                  "required": ["artifact_id", "agent_created", "file_name"]
                }
              }
            }
          },
          "404": {
            "description": "Unable to find entity with a given identifier",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "description": "Message stating the entity was not found",
                      "type": "string",
                      "example": "Unable to find entity with the provided id"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "default": { "description": "Internal Server Error" }
        },
        "tags": ["agent"]
      }
    },
    "/ap/v1/agent/tasks/{task_id}/artifacts/{artifact_id}": {
      "get": {
        "operationId": "downloadAgentTaskArtifact",
        "summary": "Download a specified artifact.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "ID of the task",
            "required": true,
            "schema": { "type": "string" },
            "example": "50da533e-3904-4401-8a07-c49adf88b5eb",
            "x-postman-variables": [{ "type": "load", "name": "task_id" }]
          },
          {
            "name": "artifact_id",
            "in": "path",
            "description": "ID of the artifact",
            "required": true,
            "schema": { "type": "string" },
            "example": "1e41533e-3904-4401-8a07-c49adf8893de"
          }
        ],
        "responses": {
          "200": {
            "description": "Returned the content of the artifact.",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "404": {
            "description": "Unable to find entity with a given identifier",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "description": "Message stating the entity was not found",
                      "type": "string",
                      "example": "Unable to find entity with the provided id"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "default": { "description": "Internal Server Error" }
        },
        "tags": ["agent"]
      }
    }
  },
...
}